extract spot vm option helper to cloud/converters
